% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_est.R
\name{naive_est}
\alias{naive_est}
\title{Computes the naive estimator for mean Time in Range (TIR).}
\usage{
naive_est(
  data,
  min_time = 0,
  max_time = (1440 * 7 - 5),
  boot = NULL,
  id_col = "patient_id",
  time = "time",
  value_in_range = "value_in_range"
)
}
\arguments{
\item{data}{A data frame containing long format time-series data.}

\item{min_time}{The minimum time threshold for filtering (default = 0).}

\item{max_time}{The maximum time threshold for filtering (default = 10075, equivalent to 1440 * 7 - 5).}

\item{boot}{Number of bootstrap iterations (default = NULL, no bootstrapping).}

\item{id_col}{Column name identifying subjects (default = "patient_id").}

\item{time}{Column name representing time (default = "time").}

\item{value_in_range}{Column name of the binary indicator (0/1) for time in range (default = "value_in_range").}
}
\value{
A list containing TIR estimation results, including the estimate, standard error, and confidence intervals.
}
\description{
This function estimates the mean TIR for a given dataset, filtering based on a
specified time window. If bootstrapping is enabled, it computes confidence
intervals using resampling.
}
