% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_in_range.R
\name{value_in_range}
\alias{value_in_range}
\title{Check if a number is within the specified range.}
\usage{
value_in_range(x, range = c(-.Machine$integer.max, .Machine$integer.max))
}
\arguments{
\item{x}{A numeric value to check.}

\item{range}{A numeric vector of length 2 specifying the lower and upper bounds.}
}
\value{
An integer: 1 if x is within the range, 0 otherwise.
}
\description{
Check if a number is within the specified range.
}
